+++ ExposurePlayback/Context/ExposureContext.swift
@@ -38,9 +38,6 @@ public class ExposureContext: MediaContext {
     /// Service that manages entitlement validation on currently active program
     public var programService: ProgramService?
     
-    /// Service that manages contract restrictions
-    internal let contractRestrictionsService: ContractRestrictionsService
-    
     /// Service that listens to changes in network connection
     internal var reachability: Reachability?
     
@@ -49,6 +46,9 @@ public class ExposureContext: MediaContext {
     
     /// Tracks the internal entitlementResponse callback
     internal var onEntitlementResponse: (PlaybackEntitlement, Source) -> Void = { _,_ in }
+    
+    /// Tracks the internal adService callback
+    internal var onAdServiceRequested: (Source) -> Void = { _ in }
     
     /// Specifies playback related behaviour
     internal(set) public var playbackProperties: PlaybackProperties = PlaybackProperties()
@@ -81,8 +81,6 @@ public class ExposureContext: MediaContext {
         self.environment = environment
         self.sessionToken = sessionToken
         self.monotonicTimeService = MonotonicTimeService(environment: environment)
-        self.contractRestrictionsService = ContractRestrictionsService()
-        
         
         reachability = Reachability()
         do {