+++ ExposurePlaybackTests/SeekToLive/ProgramSource+StaticManifest+SeekToLive.swift
@@ -67,7 +67,7 @@ class StaticProgramSourceSeekToLiveSpec: QuickSpec {
 
                     // Configure the playable
                     let provider = MockedProgramEntitlementProvider()
-                    provider.mockedRequestEntitlement = { _,_,_,_, callback in
+                    provider.mockedRequestEntitlement = { _,_,_, callback in
                         var json = PlaybackEntitlement.requiedJson
                         json["mediaLocator"] = "file://play/.isml"
                         json["playSessionId"] = "SeekToLiveTrigger"
@@ -148,7 +148,7 @@ class StaticProgramSourceSeekToLiveSpec: QuickSpec {
 
                     // Configure the playable
                     let provider = MockedProgramEntitlementProvider()
-                    provider.mockedRequestEntitlement = { _,_,_,_, callback in
+                    provider.mockedRequestEntitlement = { _,_,_, callback in
                         var json = PlaybackEntitlement.requiedJson
                         json["mediaLocator"] = "file://play/.isml"
                         json["playSessionId"] = "SeekToLiveTrigger"