+++ ExposurePlayback/Context/Source/AssetSource.swift
@@ -19,9 +19,8 @@ import Player
 /// * `.defaultBehavior` Playback starts from the beginning of the asset
 open class AssetSource: ExposureSource {
     public override func prepareSourceUrl(callback: @escaping (URL?) -> Void) {
-        if let adService = adService {
-            let sourceUrl = entitlement.adMediaLocator ?? entitlement.mediaLocator
-            adService.prepareAsset(source: sourceUrl) {
+        if let adService = adService, let adMediaLocator = entitlement.adMediaLocator {
+            adService.prepareAsset(source: adMediaLocator) {
                 callback($0)
             }
         }