+++ CHANGELOG.md
@@ -14,9 +14,17 @@
 
 ## 2.0.93
 
+#### Features
+* `EMP-11839` Introduced `AdServiceProvider` to handle server side ad insertion.
+
 #### Bug Fixes
 * `EMP-11909` Fixed inconsistent behaviour when seeking by position in a *unix time based stream*. Bounds checking and relevant callbacks should now use the correct offset.
 
+#### Changes
+* `EMP-11894` Contract restrictions service now part of source instead of context.
+
+#### Known Limitations
+* `EMP-11863` *Carthage*, using `xcodebuild` and `Xcode10`s new build system, fails to resolve and link the correct dependencies when `ExposurePlayback` is included as a dependency.
 
 ## 2.0.92