+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+* `2.0.95` Release - [2.0.95](#2095)
 * `2.0.93` Release - [2.0.93](#2093)
 * `2.0.92` Release - [2.0.92](#2092)
 * `2.0.91` Release - [2.0.91](#2091)
@@ -11,6 +12,11 @@
 * `2.0.79` Release - [2.0.79](#2079)
 * `2.0.78` Release - [2.0.78](#2078)
 * `0.77.x` Releases - [0.77.0](#0770)
+
+## 2.0.93
+
+#### Changes
+* Source preparation for ad-based playback now requires both an `adMediaLocator` and an attached `AdService` to work.
 
 ## 2.0.93